move proto conversions into their own file
